made chunksize long instead of int

+++ src/main/scala/ch/ethz/queries/chb/Q10.scala
@@ -22,7 +22,7 @@ order by revenue desc
  */
 class Q10  extends ChQuery {
 
-  override def execute(st: String, cm: String, cn:Int, cs:Int, mUrl:String): Unit = {
+  override def execute(st: String, cm: String, cn:Int, cs:Long, mUrl:String): Unit = {
     val scc = new TSparkContext(mUrl, className, st, cm, cn, cs)
     val sqlContext = new org.apache.spark.sql.SQLContext(scc.sparkContext)
     import org.apache.spark.sql.functions._
