+++ src/main/scala/ch/ethz/queries/chb/Q11.scala
@@ -28,7 +28,7 @@ having   sum(s_order_cnt) >
 order by ordercount desc
    */
 
-  override def execute(st: String, cm: String, cn:Int, cs:Int, mUrl:String): Unit = {
+  override def execute(st: String, cm: String, cn:Int, cs:Long, mUrl:String): Unit = {
     val scc = new TSparkContext(mUrl, className, st, cm, cn, cs)
     val sqlContext = new org.apache.spark.sql.SQLContext(scc.sparkContext)
